Master/requirements.txt
adafruit-blinka-bleio 
adafruit-circuitpython-ble